Fixed a critical bug
There was a bug where finding lnkfilename method contained a hardcoded value of > 51200 bytes. It was a workaround until I find a solution as it may cause bugs if there are multiple lnk files with higher value. This was fixed by storing the total size of the LNK file inside the first stage.